+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.0
+
+* Dotted animation added
+  * Feature 1: Dotted underline animated is added.
+  * Feature 2: dotRadius(optional) new parameter added.
+  * Feature 3: dotSpace(optional) new parameter added.
+
+
 ## 1.0.0
 
 * Initial release with basic functionality.
@@ -5,3 +13,5 @@
   * Feature 2: Basic Squggly underline added.
   * Feature 3: Can be made into a link.
   * Feature 4: animation duration can be customized.
+
+